more debug

+++ package-lists/update-tests/testall.sh
@@ -1,7 +1,13 @@
 for i in *.xml.in; do 
 echo $i
-./update.sh $i > list.update
+./update.sh $i > $i.update
+if test ! -s $i.update; then
+  grep -C3 === $i.output
+  : > $i.remove
+  continue
+fi
+
 sed -e 's,delete_unmaintained="false",delete_unmaintained="true",' $i > ${i/update.xml.in/remove.xml}
-./update.sh ${i/update.xml.in/remove.xml} > list.remove
-diff -u list.update list.remove | grep '^[-+]'
+./update.sh ${i/update.xml.in/remove.xml} > $i.remove
+diff -u $i.update $i.remove | grep '^[-+]'
 done
